requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-docx==0.8.11
reportlab==4.0.4
Pillow==10.0.1
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.24.3
tabula-py==2.8.2
pandas==2.0.3
openpyxl==3.1.2
urllib3
requests
